pyproject.toml
[tool.poetry]
name = "neps"
version = "0.1.0"
description = ""
authors = ["Danny Stoll", "Maciej Janowski", "Simon Schrodi"]
readme = "README.md"
license = "MIT"
packages = [
    { include = "neps" },
    { include = "neps_examples" },
]


[tool.poetry.dependencies]
python = ">=3.7.1,<3.8"
gpytorch = "^1.5.0"
ConfigSpace = "^0.4.19"
grakel = "^0.1.8"
numpy = "^1.21.1"
seaborn = "^0.11.1"
xgboost = "^1.4.2"
pathvalidate = "^2.4.1"
pandas = "^1.3.1"
networkx = "^2.6.3"
tqdm = "^4.62.3"
nltk = "^3.6.4"
tabulate = "^0.8.9"
torchmetrics = "^0.5.1"
metahyper = {git = "https://github.com/automl/metahyper.git", rev = "v0.1.2"}
path = "^16.2.0"
termcolor = "^1.1.0"
fvcore = "^0.1.5"
torchvision = "0.10.0"

[tool.poetry.dev-dependencies]
jupyter = "^1.0"
pre-commit = "^2.10"
black = "20.8b1"
isort = "^5.7"
pylint = "^2.6"
pyupgrade = "^2.29.0"
pytest = "^6.2.5"

[tool.black]
line-length = 90
target-version = ['py37']

[tool.isort]
profile = 'black'
line_length = 90

[tool.pylint.messages_control]  # Can use lists now, maybe update in future
disable = """
all
"""
enable = """
# ------------------------------------
# Spelling
# ------------------------------------
invalid-characters-in-docstring,wrong-spelling-in-comment,wrong-spelling-in-docstring,
# ------------------------------------
# Basic checks
# ------------------------------------
not-in-loop,function-redefined,continue-in-finally,abstract-class-instantiated,star-needs-assignment-target,
duplicate-argument-name,return-in-init,too-many-star-expressions,nonlocal-and-global,return-outside-function,
return-arg-in-generator,invalid-star-assignment-target,bad-reversed-sequence,nonexistent-operator,
yield-outside-function,init-is-generator,nonlocal-without-binding,lost-exception,assert-on-tuple,
dangerous-default-value,duplicate-key,useless-else-on-loop,expression-not-assigned,confusing-with-statement,
unnecessary-lambda,pointless-statement,unnecessary-pass,unreachable,eval-used,exec-used,using-constant-test,
deprecated-lambda,blacklisted-name,misplaced-comparison-constant,singleton-comparison,unneeded-not,
consider-iterating-dictionary,consider-using-enumerate,empty-docstring,unidiomatic-typecheck,
condition-evals-to-constant,
# ------------------------------------
# Async
# ------------------------------------
not-async-context-manager,yield-inside-async-function,
# ------------------------------------
# Typecheck
# ------------------------------------
invalid-unary-operand-type,unsupported-binary-operation,not-callable,redundant-keyword-arg,assignment-from-no-return,
assignment-from-none,not-context-manager,repeated-keyword,missing-kwoa,no-value-for-parameter,invalid-sequence-index,
invalid-slice-index,unexpected-keyword-arg,unsupported-membership-test,unsubscriptable-object,
# ------------------------------------
# Exceptions
# ------------------------------------
bad-except-order,catching-non-exception,bad-exception-context,notimplemented-raised,raising-bad-type,
raising-non-exception,misplaced-bare-raise,duplicate-except,nonstandard-exception,binary-op-exception,
bare-except,
# ------------------------------------
# Stdlib
# ------------------------------------
bad-open-mode,redundant-unittest-assert,boolean-datetime,deprecated-method
# ------------------------------------
# Imports
# ------------------------------------
import-error,import-self,reimported,relative-import,deprecated-module,wildcard-import,misplaced-future,cyclic-import,
wrong-import-position,ungrouped-imports,multiple-imports,
# ------------------------------------
# Variables
# ------------------------------------
unpacking-non-sequence,invalid-all-object,unbalanced-tuple-unpacking,undefined-variable,undefined-all-variable,
used-before-assignment,cell-var-from-loop,global-variable-undefined,redefined-builtin,redefine-in-handler,
unused-import,unused-argument,unused-wildcard-import,unused-variable,global-variable-not-assigned,
undefined-loop-variable,global-statement,global-at-module-level,
# ------------------------------------
# Strings
# ------------------------------------
format-needs-mapping,truncated-format-string,missing-format-string-key,mixed-format-string,too-few-format-args,
bad-str-strip-call,too-many-format-args,bad-format-character,format-combined-specification,bad-format-string-key,
bad-format-string,missing-format-attribute,missing-format-argument-key,unused-format-string-argument,
unused-format-string-key,invalid-format-index,f-string-without-interpolation
# ------------------------------------
# String Constant
# ------------------------------------
anomalous-unicode-escape-in-string,anomalous-backslash-in-string,
# ------------------------------------
# Elif
# ------------------------------------
simplifiable-if-statement,
# ------------------------------------
# Logging
# ------------------------------------
logging-format-truncated,logging-too-few-args,logging-too-many-args,logging-unsupported-format,
# ------------------------------------
# Iterable
# ------------------------------------
not-an-iterable,not-a-mapping,
# -----------------------------------
# Format
# -----------------------------------
bad-indentation,unnecessary-semicolon,missing-final-newline,mixed-line-endings,multiple-statements,trailing-newlines,
trailing-whitespace,unexpected-line-ending-format,superfluous-parens,
# ------------------------------------
# Classes
# ------------------------------------
access-member-before-definition,method-hidden,assigning-non-slot,duplicate-bases,inconsistent-mro,inherit-non-class,
invalid-slots,invalid-slots-object,no-method-argument,no-self-argument,unexpected-special-method-signature,
non-iterator-returned,invalid-length-returned,protected-access,attribute-defined-outside-init,abstract-method,
bad-staticmethod-argument,non-parent-init-called,super-init-not-called,no-classmethod-decorator,
no-staticmethod-decorator,no-self-use,bad-classmethod-argument,bad-mcs-classmethod-argument,bad-mcs-method-argument,
method-check-failed,invalid-bool-returned,invalid-index-returned,invalid-repr-returned,invalid-str-returned,
invalid-bytes-returned,invalid-hash-returned,invalid-length-hint-returned,invalid-format-returned,
invalid-getnewargs-returned,invalid-getnewargs-ex-returned,super-with-arguments
"""

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
